Move stringClauseSatisfy to separate file: filter clause predicate
